+++ packages/brick-container/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.11.11](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@3.11.10...@next-core/brick-container@3.11.11) (2024-02-21)
+
+
+### Bug Fixes
+
+* fix size check: ignore alias of deprecated bricks ([ce24a2d](https://github.com/easyops-cn/next-core/commit/ce24a2d34ada18feb339d68c2b07752a50d2db6e))
+
+
+
+
+
+## [3.11.10](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@3.11.9...@next-core/brick-container@3.11.10) (2024-02-07)
+
+**Note:** Version bump only for package @next-core/brick-container
+
+
+
+
+
+## [3.11.9](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@3.11.8...@next-core/brick-container@3.11.9) (2024-02-06)
+
+**Note:** Version bump only for package @next-core/brick-container
+
+
+
+
+
+## [3.11.8](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@3.11.7...@next-core/brick-container@3.11.8) (2024-02-05)
+
+**Note:** Version bump only for package @next-core/brick-container
+
+
+
+
+
+## [3.11.7](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@3.11.6...@next-core/brick-container@3.11.7) (2024-02-04)
+
+**Note:** Version bump only for package @next-core/brick-container
+
+
+
+
+
 ## [3.11.6](https://github.com/easyops-cn/next-core/compare/@next-core/brick-container@3.11.5...@next-core/brick-container@3.11.6) (2024-02-01)
 
 **Note:** Version bump only for package @next-core/brick-container