+++ packages/supply/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.11](https://github.com/easyops-cn/next-core/compare/@next-core/supply@2.1.10...@next-core/supply@2.1.11) (2024-02-04)
+
+**Note:** Version bump only for package @next-core/supply
+
+
+
+
+
 ## [2.1.10](https://github.com/easyops-cn/next-core/compare/@next-core/supply@2.1.9...@next-core/supply@2.1.10) (2024-01-02)
 
 **Note:** Version bump only for package @next-core/supply