+++ packages/yo/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.39](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.38...@next-core/yo@1.5.39) (2024-02-07)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.38](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.37...@next-core/yo@1.5.38) (2024-02-04)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
 ## [1.5.37](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.36...@next-core/yo@1.5.37) (2024-01-31)
 
 **Note:** Version bump only for package @next-core/yo